# feat(lib-storage): improve performance by reducing buffer copies