--- a/lib/lib-storage/src/bytelength.ts
+++ b/lib/lib-storage/src/bytelength.ts
@@ -1,8 +1,11 @@
 import { ClientDefaultValues } from "./runtimeConfig";
+import { Buffer } from "buffer";
 
 export const byteLength = (input: any) => {
   if (input === null || input === undefined) return 0;
-  if (typeof input === "string") input = Buffer.from(input);
+  if (typeof input === "string") {
+    return Buffer.byteLength(input);
+  }
   if (typeof input.byteLength === "number") {
     return input.byteLength;
   } else if (typeof input.length === "number") {