--- a/lib/lib-storage/src/chunks/getDataReadable.ts
+++ b/lib/lib-storage/src/chunks/getDataReadable.ts
@@ -1,8 +1,13 @@
 import { Buffer } from "buffer";
 import { Readable } from "stream";
 
-export async function* getDataReadable(data: Readable): AsyncGenerator<Buffer> {
+export async function* getDataReadable(data: Readable): AsyncGenerator<Buffer | Uint8Array> {
   for await (const chunk of data) {
-    yield Buffer.from(chunk);
+    // check if is Buffer/Unit8Array
+    if (Buffer.isBuffer(chunk) || chunk instanceof Uint8Array) {
+      yield chunk;
+    } else {
+      yield Buffer.from(chunk);
+    }
   }
 }