--- a/packages/node-http-handler/src/server.mock.ts
+++ b/packages/node-http-handler/src/server.mock.ts
@@ -14,7 +14,7 @@ const setResponseHeaders = (response: ServerResponse, headers: HeaderBag) => {
   }
 };
 
-const setResponseBody = (response: ServerResponse, body: Readable | string) => {
+const setResponseBody = (response: ServerResponse, body: Readable | string | Buffer) => {
   if (body instanceof Readable) {
     body.pipe(response);
   } else {
@@ -60,4 +60,37 @@ export const createMockHttpServer = (): HttpServer => {
 export const createMockHttp2Server = (): Http2Server => {
   const server = createHttp2Server();
   return server;
+};
+
+export const createMirrorResponseFunction =
+  (httpResp: HttpResponse) => (request: IncomingMessage, response: ServerResponse) => {
+    const bufs: Buffer[] = [];
+    request.on("data", (chunk) => {
+      bufs.push(chunk);
+    });
+    request.on("end", () => {
+      response.statusCode = httpResp.statusCode;
+      setResponseHeaders(response, httpResp.headers);
+      setResponseBody(response, Buffer.concat(bufs));
+    });
+    request.on("error", (err) => {
+      response.statusCode = 500;
+      setResponseHeaders(response, httpResp.headers);
+      setResponseBody(response, err.message);
+    });
+  };
+
+export const getResponseBody = (response: HttpResponse) => {
+  return new Promise<string>((resolve, reject) => {
+    const bufs: Buffer[] = [];
+    response.body.on("data", function (d: Buffer) {
+      bufs.push(d);
+    });
+    response.body.on("end", function () {
+      resolve(Buffer.concat(bufs).toString());
+    });
+    response.body.on("error", (err: Error) => {
+      reject(err);
+    });
+  });
 };