--- a/packages/util-body-length-node/src/index.ts
+++ b/packages/util-body-length-node/src/index.ts
@@ -5,7 +5,7 @@ export function calculateBodyLength(body: any): number | undefined {
     return 0;
   }
   if (typeof body === "string") {
-    return Buffer.from(body).length;
+    return Buffer.byteLength(body);
   } else if (typeof body.byteLength === "number") {
     // handles Uint8Array, ArrayBuffer, Buffer, and ArrayBufferView
     return body.byteLength;