Reduce memory usage by reducing buffer-copying and returning the original buffer instead.

### Description
This PR reduces the number of buffers copied and created from existing buffers. The most significant change was made to streams that return buffers and Uint8Arrays, but some minor improvements were made for `string` and `Uint8Array` bodies as well. The following changes were made:

- Yield without copying the given Buffer/Uint8Array in `getDataReadable.ts` and `getDataReadableStream.ts` (probably the most significant change).
- In `byteLength.ts`, use `Buffer.byteLength(string)` instead of `Buffer.from(string).byteLength` for computing the string byteLength, so that the string is not copied into a buffer just for its byteLength.
- Treat `Uint8Array` the same as a Buffer in `chunker.ts`
- Use `subarray` instead of `slice` in `getChunkBuffer.ts` so that `Uint8Array` and `Buffer` inputs behave the same. In Node `.slice` on Uint8Array copies the data, while `.subarray` just takes the subarray without copying. Slice and subarray for Buffers in Node are the same (both don't copy the data).
- `getChunkStream.ts` was changed in a similar way to `getChunkBuffer.ts`, and also added a change so that the last  buffer is not copied if there's only one buffer left (the last change is probably relatively minor).

### Testing
I ran the following file for different values:

```ts
import { getChunk } from './chunker';
import { Readable } from 'stream';
const buf = Buffer.from('#'.repeat(1024 * 1024 * 10));
const getIter = () => getChunk(Readable.from((async function* () {
    const chunkSize = 1024 * 1024;
    for (let i = 0; i < buf.byteLength; i += chunkSize) {
        yield buf.subarray(i, i + chunkSize);
    }
})()), 1024 * 1024 * 5)

async function iterateTimes(times: number) {
    console.time('timing')
    for (let i = 0; i < times; i++) {
        for await (const z of getIter()) { }
    }
    console.log(`times ${times}`);
    const used = process.memoryUsage() as any;
    for (let key in used) {
        console.log(`${key} ${Math.round(used[key] / 1024 / 1024 * 100) / 100} MB`);
    }
    console.timeEnd('timing')
}

iterateTimes(10);
```
<details>
<summary>results</summary>
Original code

>>>>>
times 1
rss 59.67 MB
heapTotal 15.68 MB
heapUsed 12.78 MB
external 30.91 MB
arrayBuffers 30.02 MB
timing: 17.289ms

times 2
rss 79.71 MB
heapTotal 15.68 MB
heapUsed 12.82 MB
external 50.91 MB
arrayBuffers 50.02 MB
timing: 26.743ms

times 10
rss 173.15 MB
heapTotal 7.18 MB
heapUsed 2.3 MB
external 90.91 MB
arrayBuffers 90.02 MB
timing: 90.863ms

>>>>>
New:
times 1
rss 49.52 MB
heapTotal 15.68 MB
heapUsed 12.78 MB
external 20.91 MB
arrayBuffers 20.02 MB
timing: 13.046ms

times 2
rss 59.64 MB
heapTotal 15.43 MB
heapUsed 12.8 MB
external 30.91 MB
arrayBuffers 30.02 MB
timing: 17.994ms

times 10
rss 129.96 MB
heapTotal 7.18 MB
heapUsed 2.32 MB
external 60.91 MB
arrayBuffers 60.02 MB
timing: 59.366ms
</details>

Note that I executed the above multiple times, to see that the results are essentially the same, but I didn't do a z-test or anything. If there's any benchmarking tools that I can use to have better results to show that the results are significant, I'd be happy to use them. From my testing, the new chunking was faster and also used less memory, when chunking a stream of a 10MB buffer with 1MB chunks. I also checked to see if there was some degraded performance for streams that return strings instead of buffers, but performance was essentially the same.

For byteLength, I executed a similar script that called `byteLength` 10 times with a 10MB and 1MB strings and also saw significant memory reduction.
<details>
 <summary>results</summary>
10MB strings:
>>>>>
original code

times 10
rss 129.17 MB
heapTotal 14.93 MB
heapUsed 12.1 MB
external 40.93 MB
arrayBuffers 40.01 MB
timing: 90.462ms
>>>>>
new code
times 10
rss 29.49 MB
heapTotal 15.43 MB
heapUsed 12.66 MB
external 0.93 MB
arrayBuffers 0.01 MB
timing: 26.115ms

1MB strings
>>>>>
original code
times 10
rss 30.39 MB
heapTotal 6.68 MB
heapUsed 3.66 MB
external 0.93 MB
arrayBuffers 0.01 MB
timing: 13.741ms
>>>>>>>
new code
times 10
rss 20.36 MB
heapTotal 6.68 MB
heapUsed 3.67 MB
external 0.93 MB
arrayBuffers 0.01 MB
timing: 7.668ms
</details>

I did not check `Uint8Array` buffers or streams (i.e. streams that return such chunks), but I expect similar performance number improvements as above. If it's necessary I'll also check as well.

### Additional context
~If the general direction of this PR is OK, I'd be happy to add some more (regression) tests for Uint8Array buffers and streams and some other edge-cases, as those currently don't exist so if I did mess something up with such inputs, there are no tests to show it.~ Also, if there's some benchmark tooling that I missed, I'd be happy to add benchmarks.

---
By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.