@Linkgoron 2021-07-26
I've added two extra changes. They're technically not from lib-storage, but I felt that they're of the same spirit as my other changes so I thought that it made sense to put them here.

- in `util-body-length-node/src/index.ts` - don't copy the buffer for its size, just return Buffer.byteLength(string) which avoids the copy
- in `write-request-body.ts`, from what I've read https://github.com/aws/aws-sdk-js-v3/issues/1433, `string` and `Buffer` are fine as inputs for `end`, so there's no reason to copy them before sending. This change is a bit more delicate though, because it does change how strings are fed.