@Linkgoron 2021-08-07
I've Added an additional tweak for `Uint8Array`s in `node-http-handler` to reuse their memory as well, and I've added tests to verify my changes (to all of the components). The test changes look a bit larger than they actually are, as most of the changes are just whitespace changes to reuse the tests for both buffers and Uint8arrays.