@uncvrd 2022-12-07
@Linkgoron hey! I've been dealing with OOM errors when bulk uploading using the `lib-storage` package (I've reduced my code to load a stream in from s3 and then reupload it using lib-storage as a test case). Since it doesn't look like this update to reduce memory footprint will be implemented any time soon, I was curious what alternative solutions you have taken? I've really struggled to find a stream only approach that doesnt utilize buffers and ISNT 6+ year old solution haha. Would love some guidance if possible?

I assume you were running in to memory issues when using this library as well? Honestly shocked I don't see more issues pop up about the stead memory increase over time until bulk uploading is complete.

I wrote an issue here where I assume it has something to do with backpressure or the buffers not releasing from gc quick enough, but honestly I'm not experienced in this realm of node at all.

Here's my issue ticket: https://github.com/aws/aws-sdk-js-v3/issues/4257