@uncvrd 2022-12-07
> > @Linkgoron hey! I've been dealing with OOM errors when bulk uploading using the `lib-storage` package (I've reduced my code to load a stream in from s3 and then reupload it using lib-storage as a test case). Since it doesn't look like this update to reduce memory footprint will be implemented any time soon, I was curious what alternative solutions you have taken? I've really struggled to find a stream only approach that doesnt utilize buffers and ISNT 6+ year old solution haha. Would love some guidance if possible?
> > I assume you were running in to memory issues when using this library as well? Honestly shocked I don't see more issues pop up about the steady memory increase over time until bulk uploading is complete.
> > I wrote an issue here where I assume it has something to do with backpressure or the buffers not releasing from gc quick enough, but honestly I'm not experienced in this realm of node at all.
> > Here's my issue ticket: #4257
> 
> I didn't actually have any memory issues, but when i created a different PR I just noticed how many copies there were - this PR is a bit old now, so I might not remember correctly, but I believe that every buffer is copied at least twice and in some cases three times. I'd be happy to put this PR back into a mergable state if the team has interest in it.

Got it - yea turns out my issue was not due to a leak but the fact that this package does take up around 500MB before stabilizing if bulk uploading, so because my staging sites were using 512MB, this was enough for it to look like the memory was constantly increasing thus crashing my server. Increasing the memory to >512MB solved that. 

<img width="687" alt="image(1)" src="https://user-images.githubusercontent.com/25097099/206113874-460b34f8-2891-44c6-9366-d1e21f7f37a2.png">

I'm trying to bring more attention to your PR as it would be great to update this and eventually have it merged!!