@kuhe 2023-08-11
due to package migration, moved the work into new PRs

https://github.com/aws/aws-sdk-js-v3/pull/5078
https://github.com/awslabs/smithy-typescript/pull/867